Renamed 'styles.css' to 'theme.css' and added 'layout.css'
Based off previous projects I want to get a head start on keeping things organized and clean.  Therefore, it seemed like it might be worth trying to break up the css for theme and layout into different files.  As well, as I go along I will create sections in the CSS where the thing being layed out or styled will be clear and easy to find